requirements.txt
aws-cdk-lib==2.200.1
aws-cdk.asset-awscli-v1==2.2.237
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.aws-lambda-python-alpha==2.200.1a0
aws-cdk.cloud-assembly-schema==44.2.0
constructs>=10.0.0,<11.0.0

boto3~=1.38.34